% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{BamScaleR-package}
\alias{BamScaleR}
\alias{BamScaleR-package}
\title{BamScaleR: Bioconductor-Friendly Multithreaded BAM Processing}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Multithreaded sequential BAM processing built on top of the ompBAM C++ engine. BamScaleR provides user-friendly BAM read and scan interfaces designed for compatibility with existing Bioconductor workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cparsania.github.io/BamScaleR/}
  \item Report bugs at \url{https://github.com/cparsania/BamScaleR/issues}
}

}
\author{
\strong{Maintainer}: Chirag Parsania \email{chirag.parsania@gmail.com}

Authors:
\itemize{
  \item Chirag Parsania \email{chirag.parsania@gmail.com}
}

}
\keyword{internal}
